update compiler
The data-items were good as temporary result of the parser, but its way to
generic to be good handable in the hole project. So a more explizit
data-structure was implemented. The new added step of the compiler now
converts the generic data-items, which coming from the parser, into
this explizit structure. The data-items are still used there, but only
for single lists or maps. This makes it much easier to keep control over
the code, because I had already problems at some points. So the code is
now more scalable und better understandable.